Add an option to allow marshaling empty Go disjunctions
What

This change adds a new config option for Go JSON marshaling Jenny, which
makes generated code ignore empty disjunctions when marshaling.

Why

There is a use-case for it in `grafana/grafana` where returning an error
from the marshaler causes the API to panic, due to how schema validation
is enforced on the server.

Closes https://github.com/grafana/cog/issues/778

Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>